refactor - move base in quotes
