+++ quotes/src/main/java/com/example/quotes/base/ui/BaseViewModel.kt
@@ -1,11 +1,11 @@
-package com.alphanication.rebirth.ui.base
+package com.example.quotes.base.ui
 
 import androidx.lifecycle.ViewModel
 import io.reactivex.rxjava3.disposables.CompositeDisposable
 
 abstract class BaseViewModel : ViewModel() {
 
-    protected var disposables: CompositeDisposable = CompositeDisposable()
+    private var disposables: CompositeDisposable = CompositeDisposable()
 
     override fun onCleared() {
         disposables.apply {
