+++ quotes/src/main/java/com/example/quotes/ui/mainquote/MainQuoteFragment.kt
@@ -4,11 +4,9 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.fragment.app.Fragment
+import com.example.quotes.base.ui.BaseFragment
 
-//import com.alphanication.rebirth.ui.base.BaseFragment
-
-class MainQuoteFragment : Fragment() {
+class MainQuoteFragment : BaseFragment() {
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
